Plone template: combine Plone 5 resource bundle aggregations.
Plone 5 has resource bundle aggregations which should be updated when
installing a release.
The update_plone script is adapted so that it rebuilds the aggregation
after installing the update.

An error is printed when we are either not using Plone 5 or not having
an upto date ftw.ugprade version. The error will not abort the update.